backend/requirements.txt
# --- Core ---
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
PyJWT==2.8.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-multipart
python-dateutil==2.8.2
requests>=2.31.0
python-socketio>=5.10.0
python-engineio>=4.8.0

# --- ML Libraries ---
numpy==1.26.3
opencv-python-headless==4.9.0.80
insightface==0.7.3
onnxruntime>=1.17.0
pandas==2.2.3
openpyxl==3.1.2
Pillow>=10.1.0

# --- Database ---
psycopg2-binary==2.9.9

# --- Telegram Bot ---
python-telegram-bot==21.3

# --- Optional: GPU acceleration ---
# onnxruntime-gpu>=1.17.0

# --- Optional: AWS S3 support ---
# boto3==1.34.0
